Bill service and main logic completed

+++ Garage-Management-Application/src/main/java/com/danny/Garage/Management/Application/dto/SparePartDTO.java
@@ -20,7 +20,7 @@ public class SparePartDTO {
 
     private Double partNumber;
 
-    private BigDecimal partPrice;
+    private Double partPrice;
 
     private String manufacture;
